+++ changelog.md
@@ -1,2 +1,2 @@
 # 1.0.0
-- Initial release
+- Initial release (tests for Geode v4.6.2 and v4.6.3)